requirements.txt
altair==5.0.1
anyio==3.7.1
attrs==23.1.0
bardapi==0.1.29
beautifulsoup4==4.12.2
blinker==1.6.2
browser-cookie3==0.19.1
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
deep-translator==1.11.4
gitdb==4.0.10
GitPython==3.1.32
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
grpcio-status==1.48.2
grpcio==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema-specifications==2023.7.1
jsonschema==4.18.4
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0

protobuf==3.20.3
pyarrow==12.0.1
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycryptodomex==3.18.0
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
pytz==2023.3
referencing==0.30.0
requests==2.31.0
rich==13.5.1
rpds-py==0.9.2
rsa==4.9
setuptools==67.7.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
wheel==0.40.0
zipp==3.16.2
